+++ front-end/mobile/lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:mobile/pages/Welcome%20Page/welcome_screen.dart';
+import 'package:mobile/pages/WelcomePage/welcome_screen.dart';
 import 'firebase_options.dart';
 
 void main() async{
