+++ front-end/mobile/lib/pages/WelcomePage/reg_screen.dart
