+++ front-end/mobile/lib/pages/WelcomePage/typewriter_text.dart
